rgw: move datalog trimming into rgw_trim_datalog.cc
